ODIN(offline)/requirements.txt
torch
torchvision
Pillow
moondream
pyttsx3
groq
matplotlib
